Move SDL2 controller bridge to utils/sim/ and add documentation
Extract the SDL2 bridge from physics.py into its own module at
utils/sim/sdl2_controller_bridge.py with comprehensive docstring
explaining the problem, solution, and usage requirements.

Update physics.py module docstring to document both macOS
workarounds (CANcoder sim frames and controller input).
