+++ physics.py
@@ -10,21 +10,25 @@ Robot field pose is integrated from swerve kinematics and fed back to:
   - The Field2d widget (visible in the sim GUI)
   - The NavX gyro sim device (so heading-based field-relative drive works)
 
-On macOS the WPILib sim GUI (GLFW) cannot read Xbox/PlayStation controller
-inputs because Apple's Game Controller framework intercepts them.  An optional
-SDL2-based bridge detects this at runtime — when a joystick port is mapped in
-the sim GUI but reports 0 axes, the bridge takes over and feeds controller
-data into the HAL sim.  Install ``pysdl2`` and ``pysdl2-dll`` to enable
-(not needed on Windows/Linux).
+macOS notes
+-----------
+Two platform-specific workarounds are applied during simulation on Mac:
+
+1. **CANcoder sim frames unavailable at init** — CTRE phoenix6's sim backend
+   on macOS does not deliver CAN frames in time for module initialization.
+   ``baseline_relative_encoders()`` in ``swerve_module.py`` handles this by
+   defaulting the steer baseline to 0° instead of raising.  The physics engine
+   also guards against an empty module list (``len(module_states) != 4``).
+
+2. **Xbox/PlayStation controller input** — Apple's Game Controller framework
+   intercepts these devices, making them invisible to GLFW (the input backend
+   used by the WPILib sim GUI).  An SDL2-based bridge
+   (``utils/sim/sdl2_controller_bridge``) reads controllers via SDL2 and writes
+   their state to the HAL sim.  See that module's docstring for setup details.
 """
 
-import logging
-import sys
-import threading
-import time
 import typing
 
-import hal.simulation
 import ntcore
 import rev
 import wpilib
@@ -35,150 +39,10 @@ from wpimath.geometry import Pose2d, Rotation2d, Twist2d
 from wpimath.kinematics import SwerveModuleState
 from wpimath.system.plant import DCMotor
 
+from utils.sim.sdl2_controller_bridge import start_controller_bridge
+
 if typing.TYPE_CHECKING:
     from robot import MyRobot
-
-logger = logging.getLogger("physics")
-
-
-# ---------------------------------------------------------------------------
-# macOS SDL2 controller bridge
-# ---------------------------------------------------------------------------
-
-def _start_sdl2_controller_bridge() -> None:
-    """Spawn a daemon thread that discovers game controllers via SDL2 and
-    feeds their state into the HAL sim.
-
-    The first controller found maps to WPILib joystick port 0 (driver),
-    the second to port 1 (operator).  If no controllers are detected the
-    thread exits silently and the sim GUI's virtual joystick still works.
-    """
-    try:
-        import sdl2
-    except ImportError:
-        msg = (
-            "pysdl2 not installed — Xbox controller input in sim requires "
-            "'pip install pysdl2 pysdl2-dll' on macOS"
-        )
-        wpilib.reportWarning(msg)
-        return
-
-    # --- mapping tables: SDL2 game controller → WPILib XboxController ---
-    _AXIS_MAP = {
-        sdl2.SDL_CONTROLLER_AXIS_LEFTX: 0,
-        sdl2.SDL_CONTROLLER_AXIS_LEFTY: 1,
-        sdl2.SDL_CONTROLLER_AXIS_TRIGGERLEFT: 2,
-        sdl2.SDL_CONTROLLER_AXIS_TRIGGERRIGHT: 3,
-        sdl2.SDL_CONTROLLER_AXIS_RIGHTX: 4,
-        sdl2.SDL_CONTROLLER_AXIS_RIGHTY: 5,
-    }
-    # SDL2 button → bit index for HAL button bitmask.
-    # WPILib buttons are 1-based; bit 0 = button 1 (A), bit 1 = button 2 (B), etc.
-    _BUTTON_MAP = {
-        sdl2.SDL_CONTROLLER_BUTTON_A: 0,
-        sdl2.SDL_CONTROLLER_BUTTON_B: 1,
-        sdl2.SDL_CONTROLLER_BUTTON_X: 2,
-        sdl2.SDL_CONTROLLER_BUTTON_Y: 3,
-        sdl2.SDL_CONTROLLER_BUTTON_LEFTSHOULDER: 4,
-        sdl2.SDL_CONTROLLER_BUTTON_RIGHTSHOULDER: 5,
-        sdl2.SDL_CONTROLLER_BUTTON_BACK: 6,
-        sdl2.SDL_CONTROLLER_BUTTON_START: 7,
-        sdl2.SDL_CONTROLLER_BUTTON_LEFTSTICK: 8,
-        sdl2.SDL_CONTROLLER_BUTTON_RIGHTSTICK: 9,
-    }
-    _DPAD = [
-        sdl2.SDL_CONTROLLER_BUTTON_DPAD_UP,
-        sdl2.SDL_CONTROLLER_BUTTON_DPAD_DOWN,
-        sdl2.SDL_CONTROLLER_BUTTON_DPAD_LEFT,
-        sdl2.SDL_CONTROLLER_BUTTON_DPAD_RIGHT,
-    ]
-
-    def _dpad_to_pov(up, down, left, right):
-        if up and right:
-            return 45
-        if right and down:
-            return 135
-        if down and left:
-            return 225
-        if left and up:
-            return 315
-        if up:
-            return 0
-        if right:
-            return 90
-        if down:
-            return 180
-        if left:
-            return 270
-        return -1
-
-    def _bridge_thread():
-        sdl2.SDL_SetHint(sdl2.SDL_HINT_JOYSTICK_MFI, b"1")
-        sdl2.SDL_Init(sdl2.SDL_INIT_GAMECONTROLLER)
-
-        # Allow time for controller discovery
-        for _ in range(20):
-            sdl2.SDL_PumpEvents()
-            time.sleep(0.05)
-
-        # Open all game-controller-capable devices, map to WPILib ports.
-        # First controller → port 0 (driver), second → port 1 (operator).
-        controllers = []  # list of (sdl2_gc, port)
-        for idx in range(sdl2.SDL_NumJoysticks()):
-            if not sdl2.SDL_IsGameController(idx):
-                continue
-            gc = sdl2.SDL_GameControllerOpen(idx)
-            if not gc:
-                continue
-            port = len(controllers)
-            gc_name = sdl2.SDL_GameControllerName(gc)
-            if gc_name and isinstance(gc_name, bytes):
-                gc_name = gc_name.decode()
-            wpilib.reportWarning(
-                f"SDL2 bridge: '{gc_name}' → HAL joystick port {port}. "
-                f"Do NOT assign this controller in the sim GUI System Joystick panel."
-            )
-            hal.simulation.setJoystickAxisCount(port, 6)
-            hal.simulation.setJoystickButtonCount(port, 10)
-            hal.simulation.setJoystickPOVCount(port, 1)
-            hal.simulation.setJoystickIsXbox(port, True)
-            hal.simulation.setJoystickName(port, gc_name or "SDL2 Controller")
-            controllers.append((gc, port))
-
-        if not controllers:
-            # No physical controllers — exit silently so the sim GUI's
-            # virtual joystick panel still works.
-            sdl2.SDL_Quit()
-            return
-
-        # Poll loop — ~100 Hz to keep input responsive
-        while True:
-            sdl2.SDL_PumpEvents()
-
-            for gc, port in controllers:
-                for sdl_axis, wpi_axis in _AXIS_MAP.items():
-                    raw = sdl2.SDL_GameControllerGetAxis(gc, sdl_axis)
-                    hal.simulation.setJoystickAxis(port, wpi_axis, raw / 32767.0)
-
-                buttons = 0
-                for sdl_btn, bit_index in _BUTTON_MAP.items():
-                    if sdl2.SDL_GameControllerGetButton(gc, sdl_btn):
-                        buttons |= (1 << bit_index)
-                hal.simulation.setJoystickButtonsValue(port, buttons)
-
-                dpad = [
-                    bool(sdl2.SDL_GameControllerGetButton(gc, b))
-                    for b in _DPAD
-                ]
-                hal.simulation.setJoystickPOV(port, 0, _dpad_to_pov(*dpad))
-
-            hal.simulation.notifyDriverStationNewData()
-            time.sleep(0.01)
-
-    t = threading.Thread(
-        target=_bridge_thread, daemon=True, name="sdl2-controller-bridge"
-    )
-    t.start()
 
 
 class PhysicsEngine:
@@ -262,11 +126,8 @@ class PhysicsEngine:
             # ignore errors here and leave _navx_yaw as None.
             pass
 
-        # On macOS, GLFW cannot read Xbox/PlayStation controllers at all
-        # (reports 0 axes, 0 buttons).  Start the SDL2 bridge immediately
-        # for the driver controller port.
-        if sys.platform == "darwin":
-            _start_sdl2_controller_bridge()
+        # On macOS, start the SDL2 controller bridge (see module docstring).
+        start_controller_bridge()
 
     def update_sim(self, now: float, tm_diff: float) -> None:
         """
